Improved requester for platform domains
