+++ RiotSharp/Http/Interfaces/IRequester.cs
@@ -1,4 +1,5 @@
-﻿using System.Collections.Generic;
+﻿using RiotSharp.Misc;
+using System.Collections.Generic;
 using System.Threading.Tasks;
 
 namespace RiotSharp.Http.Interfaces
@@ -9,7 +10,7 @@ namespace RiotSharp.Http.Interfaces
         /// Create a get request and send it to the server.
         /// </summary>
         /// <param name="relativeUrl"></param>
-        /// <param name="rootDomain"></param>
+        /// <param name="region"></param>
         /// <param name="addedArguments"></param>
         /// <param name="useHttps"></param>
         /// <returns>The content of the response.</returns>
@@ -17,14 +18,14 @@ namespace RiotSharp.Http.Interfaces
         /// Thrown if an Http error occurs. 
         /// Contains the Http error code and error message.
         /// </exception>
-        string CreateGetRequest(string relativeUrl, string rootDomain, List<string> addedArguments = null,
+        string CreateGetRequest(string relativeUrl, Region region, List<string> addedArguments = null,
             bool useHttps = true);
 
         /// <summary>
         /// Create a get request and send it asynchronously to the server.
         /// </summary>
         /// <param name="relativeUrl"></param>
-        /// <param name="rootDomain"></param>
+        /// <param name="region"></param>
         /// <param name="addedArguments"></param>
         /// <param name="useHttps"></param>
         /// <returns>The content of the response.</returns>
@@ -32,7 +33,7 @@ namespace RiotSharp.Http.Interfaces
         /// Thrown if an Http error occurs. 
         /// Contains the Http error code and error message.
         /// </exception>
-        Task<string> CreateGetRequestAsync(string relativeUrl, string rootDomain,
+        Task<string> CreateGetRequestAsync(string relativeUrl, Region region,
             List<string> addedArguments = null, bool useHttps = true);
     }
 }
